requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
capslock==1.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
Django==3.1.5
django-allauth==0.42.0
django-celery-beat==2.1.0
django-celery-results==2.0.1
django-cors-headers==3.6.0
django-rest-auth==0.9.5
djangorestframework==3.11.2
dj-paddle
drf-yasg == 1.20.0
idna==2.10
kiwisolver==1.3.1
kombu==5.0.2
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.0
prompt-toolkit==3.0.16
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
psycopg2-binary==2.8.6
requests==2.25.1
redis
celery==5.0.5
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Faker==6.6.2
djangorestframework-gis==0.17
django-rest-framework-social-oauth2==1.1.0
drf-yasg == 1.20.0
google-api-core==1.22.2
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-api-python-client==1.12.1
djangorestframework-simplejwt==4.4.0


